cs_interview_rag/requirements.txt
# Core dependencies for CS Interview RAG Assistant
endee>=0.1.0
sentence-transformers>=2.2.0
pypdf>=3.0.0
pydantic-settings>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
numpy>=1.21.0
requests>=2.28.0
